+++ tests/test_types.py
@@ -148,7 +148,7 @@ def test_default_validators(field, value, result):
         with pytest.raises(ValidationError):
             CheckModel(**kwargs)
     else:
-        assert CheckModel(**kwargs).values[field] == result
+        assert CheckModel(**kwargs).values()[field] == result
 
 
 class DatetimeModel(BaseModel):
@@ -348,7 +348,7 @@ def test_tuple():
     m = ListDictTupleModel(d=(1, 2, '3'))
     assert m.a is None
     assert m.d == (1, 2, '3')
-    assert m.values == {'a': None, 'b': None, 'c': None, 'd': (1, 2, '3')}
+    assert m.values() == {'a': None, 'b': None, 'c': None, 'd': (1, 2, '3')}
     assert ListDictTupleModel(d='xyz').d == ('x', 'y', 'z')
     assert ListDictTupleModel(d=(i**2 for i in range(5))).d == (0, 1, 4, 9, 16)
     with pytest.raises(ValidationError) as exc_info:
@@ -376,5 +376,5 @@ def test_set():
 
     m = SetModel(v=[1, 2, 3])
     assert m.v == {1, 2, 3}
-    assert m.values == {'v': {1, 2, 3}}
+    assert m.values() == {'v': {1, 2, 3}}
     assert SetModel(v={'a', 'b', 'c'}).v == {'a', 'b', 'c'}