+++ HISTORY.rst
@@ -3,8 +3,8 @@
 History
 -------
 
-v0.2.0 (TBC)
-............
+v0.2.0 (2017-06-07)
+...................
 * **breaking change**: ``values()`` on a model is now a method not a property,
   takes ``include`` and ``exclude`` arguments
 * allow annotation only fields to support mypy